% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read HTS data, weights, metadata, etc.}
\usage{
read_data(study, project_path)
}
\arguments{
\item{study}{character. A valid study name.}

\item{project_path}{character. The project directory.}
}
\value{
read_data returns an object of class "HTS.data". It contains the data files and weights necessary for
querying the HTS datasets, used primarily by \link[surveysummarize]{summarize_data}.

The "HTS.data" object is essentially a list of data.tables broken up by data and weights.

Accessing the data:
\itemize{
  \item \strong{household} - Household data file
  \item \strong{person} - Person data file
  \item \strong{trip} - Trip data file
  \item \strong{vehicle} - Vehicle data file
}

Accessing the weights:
\itemize{
  \item \strong{household} - Household weight file. Used for weighting household and vehicle data.
  \item \strong{person} - Person weight file. Also includes trip weights at the person level.
}
}
\description{
Read HTS data, weights, metadata, etc.
}
\details{
\code{read_data} is a wrapper for reading in variables from the correct csvs
and merging the corresponding tables on correct keys. \link[data.table]{fread} and
\link[data.table]{merge} are used for performance benefits.
}
\examples{
\donttest{
# Read the data
hts_data <- read_data(
  study = 'nirpc_2018',
  project_path = 'C:/2018 NIRPC Household Travel Survey'
)

# Access the data
hts_data$data$household     # household data
hts_data$data$person        # person data
hts_data$data$trip          # trip data
hts_data$data$vehicle       # vehicle data

# Access the weights
hts_data$weights$household  # household weights
hts_data$weights$person     # person and trip weights
}

}
